refactor: Add Provider component to main.jsx for Redux store
Update the main.jsx file to import the Provider component from "react-redux" and wrap the existing QueryClientProvider component with it. This change allows the Redux store to be accessed by the components in the application. By providing the store prop to the Provider component and wrapping it around the QueryClientProvider component, the application can now access the Redux store and utilize its state management capabilities.