refactor: Update Login component to use Redux for user state
Refactor the Login component in Login.jsx to use Redux for managing the user state. Import the useDispatch and useSelector hooks from "react-redux" and replace the local state management with the Redux store. This change allows for a centralized and consistent approach to managing the user state throughout the application.