refactor: Add deleteUser function to users service
Add the deleteUser function to the users service in users.js. This function allows for the deletion of a user by sending a DELETE request to the server with the user's email and authentication token. The function handles the response and returns the data if successful. If an error occurs, it throws an error with the corresponding message.