refactor: Add logout functionality to user actions
Add the logout function to the user actions in user.js. This function dispatches the resetUserInfo action from the userReducer to reset the user's information in the Redux store and removes the "account" key from the localStorage. This change allows for proper handling of user logout and ensures that the user's information is cleared from both the client-side and the browser's storage.