refactor: Add updateUserProfile function to users service
Add the updateUserProfile function to the users service in users.js. This function allows for updating the user's profile information by sending a PUT request to the server with the updated data. The function handles the response and returns the data if successful. If an error occurs, it throws an error with the corresponding message.